+++ plugins/liboggedit/oggedit_utils.c
@@ -33,9 +33,9 @@
 #include <ctype.h>
 #include "oggedit_internal.h"
 
-uint8_t *oggedit_vorbis_channel_map(const int channel_count)
+uint8_t *oggedit_vorbis_channel_map(const unsigned channel_count)
 {
-    size_t map_size = channel_count * sizeof(uint8_t);
+    unsigned map_size = channel_count * sizeof(uint8_t);
     uint8_t *map = malloc(map_size);
     if (!map)
         return NULL;
@@ -121,7 +121,7 @@ const char *oggedit_map_tag(char *key, const char *in_or_out)
 
     /* Upper-case all Vorbis Comment tag names */
     if (*in_or_out == 'm')
-        for (size_t i = 0; key[i]; i++)
+        for (unsigned i = 0; key[i]; i++)
             key[i] = toupper(key[i]);
 
     return key;
